% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregate_tables_by_plot.R
\name{aggregate_tables_by_plot}
\alias{aggregate_tables_by_plot}
\title{Fonction d'agrégation des résultats par placettes}
\usage{
aggregate_tables_by_plot(df = NULL, group_var = NULL, vars = NULL)
}
\arguments{
\item{df}{= table d'inventaire d'entrée}

\item{group_var}{= groupe à utiliser pour agréger les données}

\item{vars}{= variagbles à agréger}
}
\value{
La fonction retourne une table d'agrégation des résultats
}
\description{
La fonction permet d'agréger les résultats d'inventaire par placette
}
