% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/agreg_by_stand.R
\name{agreg_by_stand}
\alias{agreg_by_stand}
\title{Fonction d'agrégation des résultats par ensemble}
\usage{
agreg_by_stand(
  wd = NULL,
  output_dir = NULL,
  combination_table = NULL,
  forest = NULL,
  last_cycle = NULL,
  complete_progress = NULL,
  i18n = NULL
)
}
\arguments{
\item{wd}{= répertoire de travail}

\item{output_dir}{= répertoire de sortie}

\item{combination_table}{= table listant les combinaisons de variables à suivre pour agréger les résultats}

\item{forest}{= numéro et nom de la forêt}

\item{last_cycle}{= numéro du dernier passage en inventaire}

\item{complete_progress}{= barre de progression (application shiny)}

\item{i18n}{= fonction de traduction  (application shiny)}
}
\value{
La fonction crée une liste de tables de résultats par ensemble ("results_by_group") enregistrée dans l'archive "gfTablesElaborees.Rdata"
}
\description{
La fonction permet d'agréger les résultats à différentes échelles (forêt,
}
