% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_combination_table.R
\name{build_combination_table}
\alias{build_combination_table}
\title{Fonction de combinaison des groupes de variables à analyser}
\usage{
build_combination_table(vecteur = NULL)
}
\arguments{
\item{vecteur}{= liste des bnoms de tables de résultats}
}
\value{
La fonction renvoie une table de combinaisons de variables par population
}
\description{
La fonction utilise une liste de noms de tables de résultats (nomenclature spécifique : "gfForetRege_Essence"). 
Elle décompose chaque nom de table pour retrouver les variables qui le composent. Ces éléments sont ensuite répartis dans une table
qui liste les combinaisons de variables possibles. Cette table de combinaisons est utilisée lors de l'étape d'agrégation des résultats.
Les tables de résultats obtenus en sortie de l'agrégation correspondent à la liste de noms de tables en entrée.
}
