% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_increments.R
\name{calculate_increments}
\alias{calculate_increments}
\title{Fonction de calcul des accroissements}
\usage{
calculate_increments(
  df = NULL,
  echant_change = F,
  cycles_table = NULL,
  plot_table = NULL,
  vars = c("Diam", "Gha", "Vha", "VcHa")
)
}
\arguments{
\item{df}{= table d'inventaire en entrée}

\item{echant_change}{= cet argument indique s'il y a eu un changement de protocole entre 2 inventaires}

\item{cycles_table}{= tables listant les cycles d'inventaire}

\item{plot_table}{= tables listant les placettes d'inventaire}

\item{vars}{= variables de résultats}
}
\value{
La fonction renvoie la table d'entrée en ayant rajouté les variables de résultats correspondant au calcul d'accroissement (en diamètre, en volume, ...)
}
\description{
La fonction utilise une liste de noms de tables de résultats (nomenclature spécifique : "gfForetRege_Essence"). 
Elle décompose chaque nom de table pour retrouver les variables qui le composent. Ces éléments sont ensuite répartis dans une table
qui liste les combinaisons de variables possibles. Cette table de combinaisons est utilisée lors de l'étape d'agrégation des résultats.
Les tables de résultats obtenus en sortie de l'agrégation correspondent à la liste de noms de tables en entrée.
}
