% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculs_BM_carbone.R
\name{calculs_BM_carbone}
\alias{calculs_BM_carbone}
\title{Fonction de calcul du carbone dans le bois mort au sol}
\usage{
calculs_BM_carbone(
  df = NULL,
  living_trees_table = NULL,
  species_table = NULL,
  type = NULL,
  dead_wood_carbon_content = NULL,
  dead_wood_density = NULL,
  decomposition_stage_code = NULL
)
}
\arguments{
\item{df}{= table d'inventaire en entrée}

\item{living_trees_table}{= table contenant les données de bois vivant (utilisée pour déterminer la liste des essences majortaires),}

\item{species_table}{= table des essences}

\item{type}{= type de bois mort ("Bois mort au sol" ou "Bois mort sur pied")}

\item{dead_wood_carbon_content}{= table de données carbone}

\item{dead_wood_density}{= table de données sur la densité de bois par essence}

\item{decomposition_stage_code}{= table listant les stades de décomposition rencontrés dans l'inventaire}
}
\value{
La fonction retourne la table d'entrée avec les variables de résultats relatives au stock de carbone.
}
\description{
La fonction permet de calculer les variables de résultats relatives au stock de carbone dans le bois mort.
}
