% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculs_bm_lineaire.R
\name{calculs_bm_lineaire}
\alias{calculs_bm_lineaire}
\title{Fonction de traitement du bois mort au sol < 30 cm}
\usage{
calculs_bm_lineaire(df = NULL, code_essreg = NULL)
}
\arguments{
\item{df}{= table d'inventaire en entrée}

\item{code_essreg}{= table listant les regroupements d'essence}
}
\value{
La fonction retourne la table d'entrée avec les variables de résultats.
}
\description{
La fonction permet de traiter les données d'inventaire de bois mort au sol inférieur à 30 cm de diamètre
}
