% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculs_bmp.R
\name{calculs_bmp}
\alias{calculs_bmp}
\title{Fonction de calcul des volumes de bois mort sur pied}
\usage{
calculs_bmp(df = NULL)
}
\arguments{
\item{df}{= table d'inventaire en entrée}
}
\value{
La fonction retourne la table d'entrée avec les variables de résultats.
}
\description{
La fonction permet de calculer les variables de résultats du bois mort sur pied.
}
