% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculs_bms_sup30.R
\name{calculs_bms_sup30}
\alias{calculs_bms_sup30}
\title{Fonction de calcul des volumes de bois mort au sol > 30 cm}
\usage{
calculs_bms_sup30(df = NULL, code_essreg = NULL)
}
\arguments{
\item{df}{= table d'inventaire en entrée}

\item{code_essreg}{= table listant les regroupements d'essence}
}
\value{
La fonction retourne la table d'entrée avec les variables de résultats.
}
\description{
La fonction permet de calculer les variables de résultats des bois mort au sol > 30 cm .
}
