% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculs_Carbone.R
\name{calculs_Carbone}
\alias{calculs_Carbone}
\title{Fonction de calcul des quantités de carbone pour les arbres vivants}
\usage{
calculs_Carbone(
  df = NULL,
  height_by_class_and_NumTarif = NULL,
  DecroissanceRx = NULL,
  yield_table = NULL,
  DureeVieD = NULL,
  DureeVieConnexes = NULL
)
}
\arguments{
\item{df}{= table d'inventaire en entrée}

\item{height_by_class_and_NumTarif}{= à préciser}

\item{DecroissanceRx}{= à préciser}

\item{yield_table}{= table de rendements - à préciser}

\item{DureeVieD}{= à préciser}

\item{DureeVieConnexes}{= à préciser}
}
\value{
La fonction retourne la table d'entrée avec les variables de résultats.
}
\description{
La fonction permet de calculer les variables de résultats relatives au stock de carbone dans le bois vivant sur pied.
}
