% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculs_cercles.R
\name{calculs_cercles}
\alias{calculs_cercles}
\title{Fonction de calcul des résultats sur le cercle}
\usage{
calculs_cercles(
  df = NULL,
  population = NULL,
  dist_max = NULL,
  add_bmp_vars = F,
  code_essreg = NULL,
  diam_cat = NULL
)
}
\arguments{
\item{df}{= table d'inventaire en entrée}

\item{population}{= définition de la population concernée}

\item{add_bmp_vars}{= variable indiquant si des variables de résultat propres au bois mort sont à rajouter}

\item{code_essreg}{= table contenant la liste des regroupement d'essences}

\item{diam_cat}{= table listant les différentes catégories de diamètre}
}
\value{
La fonction retourne la table d'entrée avec les variables de résultats.
}
\description{
La fonction permet de calculer les variables de résultats pour la/les population(s) inventoriées sur un cercle de rayon défini.
}
