% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculs_dmh.R
\name{calculs_dmh}
\alias{calculs_dmh}
\title{Fonction de traitement des données de dendromicrohabitats (DMH)}
\usage{
calculs_dmh(df = NULL, dmh_df = NULL)
}
\arguments{
\item{df}{= table d'inventaire en entrée}

\item{dmh_df}{= table de codification des DMH}
}
\value{
La fonction retourne la table "Codes" contenant las liste des DMH rencontrés pour des calculs de fréquence/densité.
}
\description{
La fonction récupère les DMH inventoriés sur chaque arbre, décompose et liste l'ensemble des DMH rencontrés dans une table "Codes".
}
