% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculs_Eco.R
\name{calculs_Eco}
\alias{calculs_Eco}
\title{Fonction de calcul des variables éconoomiques des arbres}
\usage{
calculs_Eco(df = NULL, code_prix = NULL)
}
\arguments{
\item{df}{= table d'inventaire en entrée}

\item{code_prix}{= table listant les prix par essence, classe de diamètre et qualité}
}
\value{
La fonction renvoie la table d'entrée avec les variables économiques des arbres (Vha)
}
\description{
La fonction calcule les variables économiques de chaque arbre en fonction de la mercuriale de prix choisie
}
