% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculs_Nha.R
\name{calculs_Nha}
\alias{calculs_Nha}
\title{Fonction de calcul du poids pour les arbres}
\usage{
calculs_Nha(df = NULL)
}
\arguments{
\item{df}{= table d'inventaire en entrée}
}
\value{
La fonction renvoie la table d'entrée avec la variable du poids des arbres (Nha)
}
\description{
La fonction calcule le poids de chaque arbre en fonction des paramètres 
d'échantillonnage.
}
