% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculs_rege.R
\name{calculs_rege}
\alias{calculs_rege}
\title{Fonction de traitement de la régénération}
\usage{
calculs_rege(df = NULL, code_essreg = NULL)
}
\arguments{
\item{df}{= table d'inventaire en entrée}

\item{code_essreg}{= table contenant la liste des regroupement d'essences}
}
\value{
La fonction retourne la table d'entrée avec les variables de résultats.
}
\description{
La fonction permet de traiter les données d'inventaire de régénération.
}
