% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculs_Vol.R
\name{calculs_Vol}
\alias{calculs_Vol}
\title{Fonction de calcul du volume des arbres}
\usage{
calculs_Vol(df = NULL, IFN = F, Sup = F)
}
\arguments{
\item{df}{= table d'inventaire en entrée}

\item{IFN}{= variable permettant de rajouter un volume "IFN"}

\item{Sup}{= variable permettant de calculer le volume pour la classe de diamètre au-dessus de celle de l'arbre 
(utile dans les calculs de variables économiques)}
}
\value{
La fonction renvoie la table d'entrée avec la variable du volume des arbres (Vha)
}
\description{
La fonction calcule le volume de chaque arbre en fonction des tarifs de cubage choisis
}
