% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/change_protocole.R
\name{change_protocole}
\alias{change_protocole}
\title{Traite les changements de protocole}
\usage{
change_protocole(echant_DF)
}
\arguments{
\item{echant_DF}{= table d'échantillonnage}
}
\value{
La fonction retourne une table contenant les paramètres d'échantillonnage utilisés dans les différents calculs de résultats.
}
\description{
La fonction permet de traiter les changements de méthodes d'échantillonnage entre 2 passages en 
inventaire. Elle récupère entre 2 méthodes les paramètres d'échantillonnage correspondant aux plus petits dénominateurs communs. De cette façon, il est possible de calculer des résultats d'accroissement malgré le changement de protocole.
}
