% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_duplicated_trees.R
\name{check_duplicated_trees}
\alias{check_duplicated_trees}
\title{Contôle des arbres dupliqués (arbres repérés az+dist uniquement)}
\usage{
check_duplicated_trees(table = NULL)
}
\arguments{
\item{table}{= table d'inventaire en entrée}
}
\value{
La fonction retourne la table d'entrée sans modifications. Si des doublons ont été détectés alors un message d'erreur avec un classeur listant les erreurs apparaît.
}
\description{
La fonction permet de contrôler les doublons dans la liste des arbres inventoriés.
}
