% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_sf.R
\name{check_sf}
\alias{check_sf}
\title{Vérification des données SIG (sf)}
\usage{
check_sf(sf)
}
\arguments{
\item{sf}{= simple feature en entrée}
}
\value{
La fonction retourne la table d'entrée sans modifications. Si des erreurs ou des éléments manquants sont détectés, des messages apparaissent.
}
\description{
La fonction contrôle les shapes de placettes d'inventaire et vérifie que les numéros placette et numéros de dispositif sont bien présents.
}
