% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_wb_column_names.R
\name{check_wb_column_names}
\alias{check_wb_column_names}
\title{Vérification des intitulés de colonne du classeur d'inventaire}
\usage{
check_wb_column_names(file = NULL, table_name = NULL, column_to_check = NULL)
}
\arguments{
\item{file}{= chemin du fichier d'inventaire}

\item{table_name}{= nom de la feuille d'inventaire à contrôler}

\item{column_to_check}{= liste des noms de colonne à contrôler}
}
\value{
La fonction retourne un message en cas d'erreur/éléments manquants.
}
\description{
La fonction contrôle les intitulés de colonne des feuilles du classeur d'inventaire contenant les données.
}
