% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_workbook_sheets.R
\name{check_workbook_sheets}
\alias{check_workbook_sheets}
\title{Vérification des noms de feuille du classeur d'inventaire}
\usage{
check_workbook_sheets(file = NULL, sheets_names = NULL)
}
\arguments{
\item{file}{= chemin du fichier d'inventaire}

\item{sheets_names}{= liste des noms de feuille à vérifier}
}
\value{
La fonction retourne un message en cas d'erreur/éléments manquants.
}
\description{
La fonction contrôle les intitulés des feuilles du classeur d'inventaire contenant les données.
}
