% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/choose_forest.R
\name{choose_forest}
\alias{choose_forest}
\title{Choix de la forêt à traiter}
\usage{
choose_forest(
  object_list,
  admin = Forets,
  check_all_msg = NULL,
  df_2_test = NULL
)
}
\arguments{
\item{object_list}{= liste d'objets à filtrer}

\item{admin}{= table "Forêts"}

\item{check_all_msg}{= message pour sélectionner toutes les forêts}

\item{df_2_test}{= à préciser}
}
\value{
La fonction retourne une liste de forêt message en cas d'erreur/éléments manquants.
}
\description{
La fonction permet de choisir un/des massif(s) forestier(s) à traiter parmi ceux en mémoire.
}
