% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_names.R
\name{clean_after_knit}
\alias{clean_after_knit}
\title{Fonction de nettoyage après compilation}
\usage{
clean_after_knit(output)
}
\arguments{
\item{output}{= nom du fichier .tex}
}
\description{
La fonction supprime les fichiers "superflus" dans le dossier "out" après l'édition d'un livret de résutlats
}
