% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annexes.R, R/clean_names.R
\name{clean_names}
\alias{clean_names}
\title{Fonction de nettoyage d'une chaîne de caractères}
\usage{
clean_names(string)

clean_names(string)
}
\arguments{
\item{string}{= chaîne de caractères en entrée}
}
\value{
La fonction renvoie une chaîne de caractères

La fonction renvoie une chaîne de caractères
}
\description{
La fonction permet de débarasser une chaîne de caractères (comme un nom de massif forestier, de GF, etc.) des caractères pouvant gêner le traitement des données (espaces, caractères spéciaux, ...).

La fonction permet de débarrasser une chaîne de caractères (comme un nom de massif forestier, de GF, etc.) des caractères pouvant gêner le traitement des données (espaces, caractères spéciaux, ...).
}
