% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Coords2SHP.R
\name{Coords2SHP}
\alias{Coords2SHP}
\title{Conversion des coordonnées de placettes au format shape}
\usage{
Coords2SHP(Coords = NULL)
}
\arguments{
\item{Coords}{= tables des coordonnées de placette}
}
\description{
La fonction permet de passer de coordonnées de placette enregistrées dans une table à un shape.
}
