% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_names.R
\name{create_null}
\alias{create_null}
\title{Création d'objets nuls (package)}
\usage{
create_null(objects_list)
}
\arguments{
\item{objects_list}{= liste d'objets}
}
\description{
La fonction crée une liste d'objets nuls
}
