% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dmh_split.R
\name{dmh_split}
\alias{dmh_split}
\title{Décomposition des codes de DMH}
\usage{
dmh_split(df0 = NULL, list = list)
}
\arguments{
\item{df0}{= table d'inventaire en entrée - à préciser}

\item{list}{= liste des codes DMH rencontrés - à préciser}
}
\description{
La fonction intervient dans la fonction calculs_dmh et permet de créer la table Codes
}
