% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annexes.R
\name{filter_by_forest}
\alias{filter_by_forest}
\title{Fonction pour filtrer les données par forêt}
\usage{
filter_by_forest(tables = NULL, forest_list = NULL, cycle = NULL)
}
\arguments{
\item{tables}{= liste des tables à filtrer}

\item{forest_list}{= liste des forêts à utiliser dans le filtre}

\item{cycle}{= numéro du passage en inventaire}
}
\value{
La fonction renvoie toutes les tables filtrées avec la liste des forêts de l'argument 'forest_list'
}
\description{
La fonction filtre un ensemble de tables pour ne garder que les données relatives à un/des numéro(s) de forêt donné(s).
}
