% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcul_trees.R
\name{gf_Calculs}
\alias{gf_Calculs}
\title{Fonction de calcul des résultats d'inventaire par arbre}
\usage{
gf_Calculs(
  wd = NULL,
  output_dir = NULL,
  TauxR = 0.03,
  forest = NULL,
  last_cycle = NULL,
  complete_progress = NULL,
  i18n = NULL
)
}
\arguments{
\item{wd}{= répertoire de travail}

\item{output_dir}{= répertoire de sortie}

\item{TauxR}{= taux d'actualisation}

\item{forest}{= numéro et nom de la forêt}

\item{last_cycle}{= numéro du dernier passage en inventaire}

\item{complete_progress}{= barre de progression (application shiny)}

\item{i18n}{= fonction de traduction  (application shiny)}

\item{combination_table}{= table listant les combinaisons d'attributs à utiliser pour agréger les résultats}
}
\value{
La fonction renvoie un ensemble de tables de résultats enregistrés dans une archive Rdata.
}
\description{
La fonction utilise les données d'inventaire de chaque population (arbres, régénération, bois mort, ...) 
et calcule pour chaque individu les résultats d'inventaire ramenés à l'hectare.
}
