% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/edit_remeasure_wb.R
\name{gf_ClasseurRem}
\alias{gf_ClasseurRem}
\title{Edition du classeur de saisie des remesures.}
\usage{
gf_ClasseurRem(wd = NULL, files_list = NULL, lang = "FRA")
}
\arguments{
\item{lang}{= langue (sélectionnée) de l'interface}

\item{repGF}{= répertoire contenant les données}
}
\value{
Genère un classeur excel
}
\description{
La fonction permet d'éditer un classeur excel formaté pour faciliter la saisie des remesures d'inventaire (pour le prochain passage en inventaire)
}
\author{
Valentin Demets, Bruciamacchie Max
}
