% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/edit_remeasure_fieldsheets.R
\name{gf_EditFichesRem}
\alias{gf_EditFichesRem}
\title{Edition des fiches de remesure}
\usage{
gf_EditFichesRem(
  wd = NULL,
  files_list = NULL,
  lang = "FRA",
  template = "gf_FichesRem_2021.Rnw"
)
}
\arguments{
\item{wd}{= répertoire contenant les données}

\item{files_list}{= liste des fichiers d'inventaire}

\item{lang}{= langue (sélectionnée) de l'interface}

\item{template}{= nom du template (.Rnw)}
}
\description{
La fonction permet d'éditer les fiches de remesure pour le prochain passage en inventaire (format .pdf)
}
\author{
Valentin Demets
}
