% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MnhCreate.R
\name{MnhCreate}
\alias{MnhCreate}
\title{Modèle numérique de hauteur}
\usage{
MnhCreate(mns, mnt, resol = 1)
}
\arguments{
\item{mns}{= MNS}

\item{mnt}{= MNT}

\item{resol}{= résolution (en m)}
}
\description{
création d'un MNH par différence entre un MNS et un MNT
}
\examples{
\donttest{
librarian::shelf(terra,sf,tidyverse)

}
}
\author{
Bruciamacchie Max
}
