% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StratVertical.R
\name{StratVertical}
\alias{StratVertical}
\title{Stratification verticale}
\usage{
StratVertical(r, shp, pas = 25)
}
\arguments{
\item{r}{= MNH}

\item{shp}{= unités de gestion}

\item{pas}{= résolution (entier impair)}
}
\description{
Calcul de l'indicateur stratification verticale à partir d'un MNH
}
\examples{
\donttest{

}
}
\author{
Bruciamacchie Max
}
